+++ internal/pkg/runtime/engine/singularity/container_linux.go
@@ -50,11 +50,11 @@ import (
 // - cleanup
 // - post start process
 var (
-	cryptDev      string
-	networkSetup  *network.Setup
-	cgroupManager *cgroups.Manager
-	imageDriver   image.Driver
-	umountPoints  []string
+	cryptDev       string
+	networkSetup   *network.Setup
+	imageDriver    image.Driver
+	umountPoints   []string
+	cgroupsManager cgroups.Manager
 )
 
 // defaultCNIConfPath is the default directory to CNI network configuration files.
@@ -299,10 +299,9 @@ func create(ctx context.Context, engine *EngineOperations, rpcOps *client.RPC, p
 	if os.Geteuid() == 0 && !c.userNS {
 		path := engine.EngineConfig.GetCgroupsPath()
 		if path != "" {
-			cgroupPath := filepath.Join("/singularity", strconv.Itoa(pid))
-			cgroupManager = &cgroups.Manager{Pid: pid, Path: cgroupPath}
-			if err := cgroupManager.ApplyFromFile(path); err != nil {
-				return fmt.Errorf("failed to apply cgroups resources restriction: %s", err)
+			cgroupsManager, err = cgroups.NewManagerFromFile(path, pid, "")
+			if err != nil {
+				return fmt.Errorf("while applying cgroups config: %v", err)
 			}
 		}
 	}